web-service/requirements.txt
# Requirements file (adapted for Google App Engine)
django==4.2.*  # Django LTS version (v4.2 as of 202402)
djangorestframework  # restapi support 
django-cors-headers  # 

django-environ==0.10.0
psycopg2-binary==2.9.9  
google-cloud-secret-manager==2.16.1

openai  # for OpenAI API as well as VLLM support
together  # for Together API

# pylibmc  # needs libmemcached-tools on Ubuntu; on GAE works; remarked 202409
# transformers  # AutoTokenizer
# tiktoken  # OpenAI's fast tokenizer

# served by GAE: gunicorn  # to serve the app
